lab-2/pyproject.toml
[project]
name = "lab-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "pycryptodome>=3.23.0",
]
